LayoutTests/fast/css/pseudo-invalid-002.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<link rel="stylesheet" href="../js/resources/js-test-style.css">
<script src="../js/resources/js-test-pre.js"></script>
<style>
 :invalid { background: lime; }
 :valid { background: red; }
 textarea { background: red; }
</style>
</head>
<body>
<p id="description"></p>
<form method="get">
<textarea name="victim" required></textarea>
</form>
<div id="console"></div>
<script>
description("This test performs a check for the :invalid CSS selector on a required textarea element.");

v = document.getElementsByName("victim");
for (i = 0; i < v.length; i++)
    shouldBe("document.defaultView.getComputedStyle(v[i], null).getPropertyValue('background-color')", "'rgb(0, 255, 0)'");

var successfullyParsed = true;
</script>
<script src="../js/resources/js-test-post.js"></script>
</body>
</html>
